Throw UncheckedIOException from sendRequest/sendNotification
The send request and send notifications should never fail. If they fail
it means that the underlying neovim process has crashed, which is a
mostly unrecoverable state. If the process has crashed there is no way
of restarting it from the caller site. Bubbling up an IOException makes
all of the call signatures ugly.  If the application wants to recover
they should catch the UncheckedIOException.